+++ app/pages/artisan/submissions/index.vue
@@ -1,7 +1,15 @@
 <template>
   <UDashboardPanel id="artisan-submissions">
     <template #header>
-      <UDashboardNavbar title="Colorway Submissions" />
+      <UDashboardNavbar title="Colorway Submissions">
+        <template #right>
+          <UButton
+            label="Submit a Colorway"
+            icon="hugeicons:plus-sign"
+            to="/artisan/submissions/submit"
+          />
+        </template>
+      </UDashboardNavbar>
     </template>
 
     <template #body>